extract enums and complex types form DatasetMetadata
