+++ src/main/scala/nl.knaw.dans.easy.deposit/docs/DatasetMetadata.scala
@@ -15,14 +15,10 @@
  */
 package nl.knaw.dans.easy.deposit.docs
 
-import nl.knaw.dans.easy.deposit.docs.DatasetMetadata.AccessCategory.AccessCategory
-import nl.knaw.dans.easy.deposit.docs.DatasetMetadata.DateQualifier.DateQualifier
-import nl.knaw.dans.easy.deposit.docs.DatasetMetadata.PrivacySensitiveDataPresent.PrivacySensitiveDataPresent
-import nl.knaw.dans.easy.deposit.docs.DatasetMetadata.RelationQualifier.RelationQualifier
 import nl.knaw.dans.easy.deposit.docs.DatasetMetadata._
-import nl.knaw.dans.easy.deposit.docs.JsonUtil.{ InvalidDocumentException, RichJsonInput, toJson }
-import org.joda.time.DateTime
-import org.joda.time.format.ISODateTimeFormat
+import nl.knaw.dans.easy.deposit.docs.JsonUtil.{ InvalidDocumentException, RichJsonInput }
+import nl.knaw.dans.easy.deposit.docs.dm.PrivacySensitiveDataPresent.PrivacySensitiveDataPresent
+import nl.knaw.dans.easy.deposit.docs.dm._
 import org.json4s.JsonInput
 
 import scala.util.{ Failure, Success, Try }
@@ -39,7 +35,7 @@ case class DatasetMetadata(identifiers: Option[Seq[SchemedValue[String]]] = None
                            alternativeIdentifiers: Option[Seq[SchemedValue[String]]] = None,
                            relations: Option[Seq[RelationType]] = None,
                            languagesOfFiles: Option[Seq[PossiblySchemedKeyValue[String]]] = None,
-                           dates: Option[Seq[QualifiedSchemedValue[String, DateQualifier]]] = None,
+                           dates: Option[Seq[Date]] = None,
                            sources: Option[Seq[String]] = None,
                            instructionsForReuse: Option[Seq[String]] = None,
                            publishers: Option[Seq[String]] = None,
@@ -80,63 +76,15 @@ object DatasetMetadata {
 
   val doiScheme = "id-type:DOI"
 
-  type Date = QualifiedSchemedValue[String, DateQualifier]
-
-  def Date(value: DateTime, qualifier: DateQualifier): Date = {
-    Date(value.toString(ISODateTimeFormat.date()), qualifier)
-  }
-
-  def Date(value: String, qualifier: DateQualifier): Date = QualifiedSchemedValue[String, DateQualifier](
-    Some("dcterms:W3CDTF"), value, qualifier
-  )
-
   def missingValue(label: String): InvalidDocumentException = {
     InvalidDocumentException("DatasetMetadata", new Exception(s"Please set $label"))
   }
 
-  object PrivacySensitiveDataPresent extends Enumeration {
-    type PrivacySensitiveDataPresent = Value
-    val yes, no, unspecified = PrivacySensitiveDataPresent.Value
-  }
-
-  object AccessCategory extends Enumeration {
-    type AccessCategory = Value
-    val open: AccessCategory = Value("OPEN_ACCESS")
-    val openForRegisteredUsers: AccessCategory = Value("OPEN_ACCESS_FOR_REGISTERED_USERS")
-    val restrictedGroup: AccessCategory = Value("GROUP_ACCESS")
-    val restrictedRequest: AccessCategory = Value("REQUEST_PERMISSION")
-    val otherAccess: AccessCategory = Value("NO_ACCESS")
-  }
-
-  object DateQualifier extends Enumeration {
-    type DateQualifier = Value
-    val created: DateQualifier = Value("dcterms:created")
-    val available: DateQualifier = Value("dcterms:available")
-    val date: DateQualifier = Value("dc:date")
-    val dateAccepted: DateQualifier = Value("dcterms:dateAccepted")
-    val dateCopyrighted: DateQualifier = Value("dcterms:dateCopyrighted")
-    val dateSubmitted: DateQualifier = Value("dcterms:dateSubmitted")
-    val issued: DateQualifier = Value("dcterms:issued")
-    val modified: DateQualifier = Value("dcterms:modified")
-    val valid: DateQualifier = Value("dcterms:valid")
-  }
-
-  object RelationQualifier extends Enumeration {
-    type RelationQualifier = Value
-    val hasFormat: RelationQualifier = Value("dcterms:hasFormat")
-    val hasPart: RelationQualifier = Value("dcterms:hasPart")
-    val hasVersion: RelationQualifier = Value("dcterms:hasVersion")
-    val isFormatOf: RelationQualifier = Value("dcterms:isFormatOf")
-    val isPartOf: RelationQualifier = Value("dcterms:isPartOf")
-    val isReferencedBy: RelationQualifier = Value("dcterms:isReferencedBy")
-    val isReplacedBy: RelationQualifier = Value("dcterms:isReplacedBy")
-    val isRequiredBy: RelationQualifier = Value("dcterms:isRequiredBy")
-    val isVersionOf: RelationQualifier = Value("dcterms:isVersionOf")
-    val references: RelationQualifier = Value("dcterms:references")
-    val relation: RelationQualifier = Value("dcterms:relation")
-    val replaces: RelationQualifier = Value("dcterms:replaces")
-    val requires: RelationQualifier = Value("dcterms:requires")
-    val conformsTo: RelationQualifier = Value("dcterms:conformsTo")
+  implicit class OptionalString[T](val value: Option[T]) extends AnyVal {
+    def isProvided: Boolean = value match {
+      case Some(str: String) => str.trim.nonEmpty
+      case _ => value.isDefined
+    }
   }
 
   trait RequiresNonEmpty {
@@ -149,89 +97,11 @@ object DatasetMetadata {
     }
   }
 
-  case class AccessRights(category: AccessCategory,
-                          group: Option[String],
-                         )
-
-  implicit class OptionalString[T](val value: Option[T]) extends AnyVal {
-    def isProvided: Boolean = value match {
-      case Some(str: String) => str.trim.nonEmpty
-      case _ => value.isDefined
-    }
-  }
-
-  case class Author(titles: Option[String] = None,
-                    initials: Option[String] = None,
-                    insertions: Option[String] = None,
-                    surname: Option[String] = None,
-                    role: Option[SchemedKeyValue[String]] = None,
-                    ids: Option[Seq[SchemedValue[String]]] = None,
-                    organization: Option[String] = None,
-                   ) {
-    private val hasMandatory: Boolean = organization.isProvided || (surname.isProvided && initials.isProvided)
-    private val hasRedundant: Boolean = surname.isEmpty && (titles.isProvided || insertions.isProvided)
-    private val incompleteMsg = "needs one of (organisation | surname and initials)"
-    private val redundantMsg = "without surname should have neither titles nor insertions"
-    require(hasMandatory, buildMsg(incompleteMsg))
-    require(!hasRedundant, buildMsg(redundantMsg))
-
-    private def buildMsg(s: String) = s"Author $s; got: ${ toJson(this) }"
-
-    def isRightsHolder: Boolean = role.exists(_.key == "RightsHolder")
-
-    override def toString: String = { // TODO ID's when DatasetXml implements ID's for Author fields
-      val name = Seq(titles, initials, insertions, surname)
-        .filter(_.isProvided)
-        .map(_.getOrElse(""))
-        .mkString(" ")
-      (surname.isProvided, organization.isProvided) match {
-        case (true, true) => s"$name; ${ organization.getOrElse("") }"
-        case (false, true) => organization.getOrElse("")
-        case (true, false) => name
-        case (false, false) => throw new Exception(buildMsg(incompleteMsg)) // only with wrong requires
-      }
-    }
-  }
-
-  case class SpatialPoint(scheme: String,
-                          x: Int,
-                          y: Int,
-                         ) extends RequiresNonEmpty {
-    requireNonEmptyString(scheme, "scheme")
-  }
-
-  case class SpatialBox(scheme: String,
-                        north: Int,
-                        east: Int,
-                        south: Int,
-                        west: Int,
-                       ) extends RequiresNonEmpty {
-    requireNonEmptyString(scheme, "scheme")
-  }
-
-  trait RelationType
-
-  case class Relation(qualifier: RelationQualifier,
-                      url: Option[String],
-                      title: Option[String],
-                     ) extends RequiresNonEmpty with RelationType {
-    require(title.isProvided || url.isProvided, s"Relation needs at least one of (title | url) got: ${ toJson(this) }")
-  }
-
-  case class RelatedIdentifier(scheme: Option[String],
-                               value: String,
-                               qualifier: RelationQualifier
-                              ) extends RelationType with RequiresNonEmpty {
-    requireNonEmptyString(value, "value")
-    val hasScheme: Boolean = scheme.isProvided
-  }
-
   case class QualifiedSchemedValue[S, Q](scheme: Option[S],
                                          value: String,
                                          qualifier: Q
                                         ) extends RequiresNonEmpty {
     requireNonEmptyString(value, "value")
-    val hasScheme: Boolean = scheme.isProvided
   }
 
   case class SchemedValue[S](scheme: S,
@@ -245,7 +115,6 @@ object DatasetMetadata {
                                      value: String,
                                     ) extends RequiresNonEmpty {
     requireNonEmptyString(value, "value")
-    val hasScheme: Boolean = scheme.isProvided
   }
 
   case class SchemedKeyValue[S](scheme: S,
@@ -262,8 +131,6 @@ object DatasetMetadata {
                                         value: String,
                                        ) extends RequiresNonEmpty {
     requireNonEmptyString(value, "value")
-    val hasScheme: Boolean = scheme.isProvided
-    val hasKey: Boolean = key.isProvided
   }
 }
